refactor: rename 'serialization' file to 'serde'
